+++ include/rapidjson/internal/biginteger.h
@@ -241,7 +241,7 @@ private:
         uint64_t r = 0;
         for (const char* p = begin; p != end; ++p) {
             RAPIDJSON_ASSERT(*p >= '0' && *p <= '9');
-            r = r * 10u + static_cast<unsigned>(*p - '0');
+            r = r * 10u + static_cast<unsigned>(*p - static_cast<uint64_t>('0'));
         }
         return r;
     }
